BUG: vector data must be templated by double to work in Monteverdi (Note that the image are now in float, is it good?)

+++ Code/Modules/SupervisedClassification/otbSupervisedClassificationModule2.h
@@ -58,7 +58,7 @@ public:
   /** Data typedefs */
   /// Dataset
   typedef VectorImage<float,2>         FloatingVectorImageType;
-  typedef VectorData<float,2>          VectorDataType;
+  typedef VectorData<double,2>          VectorDataType;
 
   typedef Image<int,2>                 LabelImageType;
